adding animations

+++ Shared/Views/ListView.swift
@@ -13,18 +13,24 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items){ item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
-            }.onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+        ZStack {
+            if listViewModel.items.isEmpty{
+                NoItemsView()
+            }else{
+                List{
+                    ForEach(listViewModel.items){ item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                    }.onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
+            }
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo List")
         .navigationBarItems(leading: EditButton(), trailing:
                   NavigationLink("Add", destination: AddView())
